refactore/fix: fix all PR comments
