+++ src/App.js
@@ -1,14 +1,15 @@
 import React from 'react';
+import './app.scss';
 import Header from './components/Header/Header';
-import Footer from './components/Footer/Footer';
 import PagesContainer from './components/PagesContainer/PagesContainer';
+import Footer from './components/Footer/Footer';
 
 const App = () => {
   return (
     <div className='App'>
       <Header />
       <PagesContainer />
-      {/*  <Footer />*/}
+      <Footer />
     </div>
   );
 };
