+++ src/app.scss
@@ -1,0 +1,5 @@
+.App {
+  display: flex;
+  flex-direction: column;
+  min-height: 100%;
+}
