+++ src/components/Footer/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './footer.css';
+import './footer.scss';
 
 function Footer() {
   return <div className='footer'></div>;
