+++ src/components/Footer/footer.scss
@@ -1,5 +1,5 @@
 .footer {
   width: 100%;
-  height: 4rem;
+  height: 10vh;
   background: #000;
 }
