+++ src/components/Header/header.scss
@@ -1,5 +1,5 @@
 .header {
   width: 100%;
-  height: 5rem;
+  height: 10vh;
   background: #000;
 }
