--- src/components/MembersTasks/memberTasks.css
@@ -1,62 +1,0 @@
-.memberTasksTableContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding-top: 100px;
-}
-
-.memberTasksTable {
-  width: 100%;
-  border-collapse: collapse;
-  border: 1px solid black;
-}
-
-.memberTasksTable td {
-  width: 200px;
-  height: 80px;
-  text-align: center;
-  border: 1px solid black;
-}
-
-.memberTasksTable th {
-  width: 100px;
-  height: 40px;
-  text-align: center;
-  border: 1px solid black;
-  background: rgba(9, 116, 237, 0.7);
-}
-
-.darkLine {
-  background: rgba(207, 207, 207, 0.5);
-}
-
-.userGreeting {
-  font-size: 25px;
-}
-
-.returnToFullListButton {
-  min-width: 120px;
-  min-height: 30px;
-  margin-top: 5px;
-  cursor: pointer;
-  outline: none;
-  border: 1px solid rgba(0, 140, 255, 1);
-  border-radius: 3px;
-  align-self: flex-start;
-  margin-bottom: 10px;
-}
-
-.returnToFullListButton:hover {
-  min-width: 120px;
-  min-height: 30px;
-  margin-top: 5px;
-  cursor: pointer;
-  outline: none;
-  border: 1px solid rgba(0, 140, 255, 1);
-  border-radius: 3px;
-  margin-bottom: 10px;
-  align-self: flex-start;
-  background: rgba(9, 116, 237, 1);
-  transition: background 0.5s ease, color 0.2s linear;
-}
