+++ src/components/PagesContainer/pagesContainer.scss
@@ -1,0 +1,6 @@
+.pagesContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  min-height: 80vh;
+}
