+++ src/components/Tasks/tasks.scss
@@ -1,0 +1,29 @@
+.tasksTableContainer {
+  min-height: 500px;
+  margin: 100px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: stretch;
+
+  .tasksTable {
+    width: 100%;
+    border-collapse: collapse;
+    border: 1px solid black;
+
+    & td {
+      min-width: 30px;
+      height: 80px;
+      text-align: center;
+      border: 1px solid black;
+    }
+
+    & th {
+      width: 50px;
+      height: 40px;
+      text-align: center;
+      border: 1px solid black;
+      background: rgba(9, 116, 237, 0.7);
+    }
+  }
+}
