--- src/components/common/Spinner/spiner.css
@@ -1,9 +1,0 @@
-.spinner {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  align-items: center;
-  height: 50px;
-  width: 100px;
-  margin: auto;
-}
